telegram_bot/requirements.txt
cython
wheel
python-telegram-bot[job-queue]
google-generativeai
firebase-admin
aiosqlite
langdetect
python-dotenv
